requirements.txt
fastapi
uvicorn
python-dotenv
supabase
httpx
notion-client
